Fix warning found by Pierre
/local/jolivet/petsc/src/sys/classes/random/impls/curand/curand.c: In function ‘PetscRandomGetValuesReal_CURAND’:
/local/jolivet/petsc/src/sys/classes/random/impls/curand/curand.c:24:65: warning: operand of ‘?:’ changes signedness from ‘PetscInt’ {aka ‘int’} to ‘unsigned int’ due to unsignedness of other operand [-Wsign-compare]
   24 |   size_t              nn     = n < 0 ? (size_t)(-2 * n) : n; /* handle complex case */
      |                                                           ^

Based-on-patch-by: Pierre Jolivet <pierre@joliv.et>

+++ src/sys/classes/random/impls/curand/curand.c
@@ -21,7 +21,7 @@ PETSC_INTERN PetscErrorCode PetscRandomCurandScale_Private(PetscRandom, size_t, 
 PetscErrorCode PetscRandomGetValuesReal_CURAND(PetscRandom r, PetscInt n, PetscReal *val)
 {
   PetscRandom_CURAND *curand = (PetscRandom_CURAND *)r->data;
-  size_t              nn     = n < 0 ? (size_t)(-2 * n) : n; /* handle complex case */
+  size_t              nn     = n < 0 ? (size_t)(-2 * n) : (size_t)n; /* handle complex case */
 
   PetscFunctionBegin;
 #if defined(PETSC_USE_REAL_SINGLE)
